requirements.txt
Django==3.1.12
djangorestframework==3.12.4
django-allauth==0.33.0
django-classy-tags==2.0.0
django-cors-headers==3.7.0
django_ckeditor==6.1.0
Pillow==8.2.0
tempus_dominus==5.1.2.14

